Test refactor
Refactor test projects to clean up nuget and general references

Related to #106
