+++ Qwiq.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.9
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Core.Tests", "test\Qwiq.Core.Tests\Qwiq.Core.Tests.csproj", "{57407CCA-8444-4713-95E9-CFC1168D846B}"
 EndProject
@@ -26,6 +26,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Mocks", "test\Qwiq.Mocks\Qwiq.Mocks.csproj", "{DB07E690-4B77-414F-91C7-1A48C9F01F24}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Benchmark", "test\Qwiq.Benchmark\Qwiq.Benchmark.csproj", "{D9ED32D7-03FA-468B-AD1A-249CEF9C6CDB}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Tests.Common", "test\Qwiq.Tests.Common\Qwiq.Tests.Common.csproj", "{B45C92B0-AC36-409D-86A5-5428C87384C3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Identity.Benchmark.Tests", "test\Qwiq.Identity.Benchmark.Tests\Qwiq.Identity.Benchmark.Tests.csproj", "{26DFD8E0-C486-4EAE-BA7F-277A40AF5CC2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Qwiq.Mapper.Benchmark.Tests", "test\Qwiq.Mapper.Benchmark.Tests\Qwiq.Mapper.Benchmark.Tests.csproj", "{DB2C0AC6-CF65-412E-93A3-D3F51857BED9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,6 +87,18 @@ Global
 		{D9ED32D7-03FA-468B-AD1A-249CEF9C6CDB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9ED32D7-03FA-468B-AD1A-249CEF9C6CDB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D9ED32D7-03FA-468B-AD1A-249CEF9C6CDB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B45C92B0-AC36-409D-86A5-5428C87384C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B45C92B0-AC36-409D-86A5-5428C87384C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B45C92B0-AC36-409D-86A5-5428C87384C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B45C92B0-AC36-409D-86A5-5428C87384C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26DFD8E0-C486-4EAE-BA7F-277A40AF5CC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26DFD8E0-C486-4EAE-BA7F-277A40AF5CC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26DFD8E0-C486-4EAE-BA7F-277A40AF5CC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26DFD8E0-C486-4EAE-BA7F-277A40AF5CC2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB2C0AC6-CF65-412E-93A3-D3F51857BED9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB2C0AC6-CF65-412E-93A3-D3F51857BED9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB2C0AC6-CF65-412E-93A3-D3F51857BED9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB2C0AC6-CF65-412E-93A3-D3F51857BED9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
