+++ test/Qwiq.Core.Tests/WorkItemStoreFactoryTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 
 using Microsoft.Qwiq.Credentials;
+using Microsoft.Qwiq.Tests.Common;
 using Microsoft.VisualStudio.Services.Client;
 using Microsoft.VisualStudio.Services.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
