+++ test/Qwiq.Core.Tests/WorkItemStoreTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using Microsoft.Qwiq.Core.Tests.Mocks;
 using Microsoft.Qwiq.Proxies;
+using Microsoft.Qwiq.Tests.Common;
 using Microsoft.TeamFoundation.Client;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
