+++ test/Qwiq.Identity.Benchmark.Tests/packages.config
@@ -1,0 +1,50 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="BenchmarkDotNet" version="0.9.9" targetFramework="net46" />
+  <package id="BenchmarkDotNet.Core" version="0.9.9" targetFramework="net46" />
+  <package id="BenchmarkDotNet.Toolchains.Roslyn" version="0.9.9" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.Common" version="1.3.2" targetFramework="net46" />
+  <package id="Microsoft.CodeAnalysis.CSharp" version="1.3.2" targetFramework="net46" />
+  <package id="System.AppContext" version="4.1.0" targetFramework="net46" />
+  <package id="System.Collections" version="4.0.11" targetFramework="net46" />
+  <package id="System.Collections.Concurrent" version="4.0.12" targetFramework="net46" />
+  <package id="System.Collections.Immutable" version="1.2.0" targetFramework="net46" />
+  <package id="System.Console" version="4.0.0" targetFramework="net46" />
+  <package id="System.Diagnostics.Debug" version="4.0.11" targetFramework="net46" />
+  <package id="System.Diagnostics.FileVersionInfo" version="4.0.0" targetFramework="net46" />
+  <package id="System.Diagnostics.StackTrace" version="4.0.1" targetFramework="net46" />
+  <package id="System.Diagnostics.Tools" version="4.0.1" targetFramework="net46" />
+  <package id="System.Dynamic.Runtime" version="4.0.11" targetFramework="net46" />
+  <package id="System.Globalization" version="4.0.11" targetFramework="net46" />
+  <package id="System.IO.FileSystem" version="4.0.1" targetFramework="net46" />
+  <package id="System.IO.FileSystem.Primitives" version="4.0.1" targetFramework="net46" />
+  <package id="System.Linq" version="4.1.0" targetFramework="net46" />
+  <package id="System.Linq.Expressions" version="4.1.0" targetFramework="net46" />
+  <package id="System.Reflection" version="4.1.0" targetFramework="net46" />
+  <package id="System.Reflection.Metadata" version="1.3.0" targetFramework="net46" />
+  <package id="System.Reflection.Primitives" version="4.0.1" targetFramework="net46" />
+  <package id="System.Resources.ResourceManager" version="4.0.1" targetFramework="net46" />
+  <package id="System.Runtime" version="4.1.0" targetFramework="net46" />
+  <package id="System.Runtime.Extensions" version="4.1.0" targetFramework="net46" />
+  <package id="System.Runtime.Handles" version="4.0.1" targetFramework="net46" />
+  <package id="System.Runtime.InteropServices" version="4.1.0" targetFramework="net46" />
+  <package id="System.Runtime.Numerics" version="4.0.1" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.2.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Encoding" version="4.0.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.Primitives" version="4.0.0" targetFramework="net46" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.1.0" targetFramework="net46" />
+  <package id="System.Text.Encoding" version="4.0.11" targetFramework="net46" />
+  <package id="System.Text.Encoding.CodePages" version="4.0.1" targetFramework="net46" />
+  <package id="System.Text.Encoding.Extensions" version="4.0.11" targetFramework="net46" />
+  <package id="System.Threading" version="4.0.11" targetFramework="net46" />
+  <package id="System.Threading.Tasks" version="4.0.11" targetFramework="net46" />
+  <package id="System.Threading.Tasks.Extensions" version="4.0.0" targetFramework="net46" />
+  <package id="System.Threading.Tasks.Parallel" version="4.0.1" targetFramework="net46" />
+  <package id="System.Threading.Thread" version="4.0.0" targetFramework="net46" />
+  <package id="System.Xml.ReaderWriter" version="4.0.11" targetFramework="net46" />
+  <package id="System.Xml.XDocument" version="4.0.11" targetFramework="net46" />
+  <package id="System.Xml.XmlDocument" version="4.0.1" targetFramework="net46" />
+  <package id="System.Xml.XPath" version="4.0.1" targetFramework="net46" />
+  <package id="System.Xml.XPath.XDocument" version="4.0.1" targetFramework="net46" />
+</packages>
