+++ test/Qwiq.Linq.Tests/QueryBuilderTests.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Qwiq.Core.Tests;
+
 using Microsoft.Qwiq.Linq.Visitors;
 using Microsoft.Qwiq.Mocks;
+using Microsoft.Qwiq.Tests.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Should;
