+++ test/Qwiq.Linq.Tests/packages.config
@@ -1,42 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="GitVersionTask" version="3.6.5" targetFramework="net46" developmentDependency="true" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.2" targetFramework="net46" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.2" targetFramework="net46" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="3.13.5" targetFramework="net46" />
-  <package id="Microsoft.TeamFoundation.DistributedTask.Common" version="15.112.1" targetFramework="net46" />
-  <package id="Microsoft.TeamFoundationServer.Client" version="15.112.1" targetFramework="net46" />
-  <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net46" />
-  <package id="Microsoft.VisualStudio.Services.Client" version="15.112.1" targetFramework="net46" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.2.1" targetFramework="net46" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net46" />
   <package id="Should" version="1.1.20" targetFramework="net46" />
-  <package id="System.Collections" version="4.0.11" targetFramework="net46" />
-  <package id="System.Collections.Concurrent" version="4.0.12" targetFramework="net46" />
-  <package id="System.Diagnostics.Debug" version="4.0.11" targetFramework="net46" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.0.0" targetFramework="net46" />
-  <package id="System.Globalization" version="4.0.11" targetFramework="net46" />
-  <package id="System.IdentityModel.Tokens.Jwt" version="4.0.4.403061554" targetFramework="net46" />
-  <package id="System.IO" version="4.1.0" targetFramework="net46" />
-  <package id="System.Linq" version="4.1.0" targetFramework="net46" />
-  <package id="System.Net.Http" version="4.0.0" targetFramework="net46" />
-  <package id="System.Net.Primitives" version="4.0.11" targetFramework="net46" />
-  <package id="System.Reflection" version="4.1.0" targetFramework="net46" />
-  <package id="System.Reflection.Extensions" version="4.0.1" targetFramework="net46" />
-  <package id="System.Runtime" version="4.1.0" targetFramework="net46" />
-  <package id="System.Runtime.Extensions" version="4.1.0" targetFramework="net46" />
-  <package id="System.Runtime.InteropServices" version="4.1.0" targetFramework="net46" />
-  <package id="System.Runtime.Serialization.Json" version="4.0.2" targetFramework="net46" />
-  <package id="System.Runtime.Serialization.Primitives" version="4.1.1" targetFramework="net46" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.2.0" targetFramework="net46" />
-  <package id="System.Security.Cryptography.Encoding" version="4.0.0" targetFramework="net46" />
-  <package id="System.Security.Cryptography.Primitives" version="4.0.0" targetFramework="net46" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.1.0" targetFramework="net46" />
-  <package id="System.Text.Encoding" version="4.0.11" targetFramework="net46" />
-  <package id="System.Text.Encoding.Extensions" version="4.0.11" targetFramework="net46" />
-  <package id="System.Text.RegularExpressions" version="4.1.0" targetFramework="net46" />
-  <package id="System.Threading" version="4.0.11" targetFramework="net46" />
-  <package id="System.Threading.Tasks" version="4.0.11" targetFramework="net46" />
-  <package id="System.Xml.ReaderWriter" version="4.0.11" targetFramework="net46" />
-  <package id="WindowsAzure.ServiceBus" version="3.3.2" targetFramework="net46" />
 </packages>
