+++ test/Qwiq.Mapper.Benchmark.Tests/PerformanceTests.cs
@@ -5,10 +5,9 @@ using System.Linq;
 
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
-using Microsoft.Qwiq.Core.Tests;
-using Should;
 using Microsoft.Qwiq.Mapper.Attributes;
 using Microsoft.Qwiq.Mocks;
+using Microsoft.Qwiq.Tests.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using Newtonsoft.Json;
