+++ test/Qwiq.Mocks/MockIdentityManagementService.cs
@@ -5,9 +5,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-using Microsoft.TeamFoundation.Common;
-using Microsoft.VisualStudio.Services.Common;
-
 namespace Microsoft.Qwiq.Mocks
 {
     public class MockIdentityManagementService : IIdentityManagementService
