+++ test/Qwiq.Mocks/MockWorkItem.cs
@@ -4,8 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Runtime.Serialization.Formatters.Binary;
 
-using Microsoft.TeamFoundation.WorkItemTracking.Client;
-
 namespace Microsoft.Qwiq.Mocks
 {
     [Serializable]
