+++ test/Qwiq.Relatives.Tests/RelativesAwareTeamFoundationServerWorkItemQueryProviderTests.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Qwiq.Core.Tests;
+
 using Should;
 using Microsoft.Qwiq.Linq;
 using Microsoft.Qwiq.Linq.Visitors;
@@ -11,6 +11,7 @@ using Microsoft.Qwiq.Mocks;
 using Microsoft.Qwiq.Relatives.Linq;
 using Microsoft.Qwiq.Relatives.Mapper;
 using Microsoft.Qwiq.Relatives.Tests.Mocks;
+using Microsoft.Qwiq.Tests.Common;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Qwiq.Relatives.Tests
