+++ test/Qwiq.Tests.Common/Properties/AssemblyInfo.cs
@@ -1,16 +1,15 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Qwiq.Identity.Tests")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyTitle("Microsoft.Qwiq.Tests.Common")]
+[assembly: AssemblyDescription("Common objects used for Qwiq tests")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Qwiq.Identity.Tests")]
-[assembly: AssemblyCopyright("Copyright ©  2016")]
+[assembly: AssemblyCompany("Microsoft")]
+[assembly: AssemblyProduct("Microsoft.Qwiq")]
+[assembly: AssemblyCopyright("\x00a9 Microsoft Corporation. All rights reserved.")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -20,6 +19,4 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("ce68530e-eb8f-4be2-9563-a09ac70ea8c1")]
-
-[assembly: InternalsVisibleTo("Microsoft.Qwiq.Identity.Benchmark.Tests")]
+[assembly: Guid("b45c92b0-ac36-409d-86a5-5428c87384c3")]
